requirements.txt
streamlit
ultralytics
torch
numpy
Pillow
